backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.16.0
pydantic==2.10.6
bcrypt==4.2.1
PyJWT==2.8.0
python-dotenv==1.0.1